app/requirements.txt
pytest
coverage
distlib
filelock
packaging
platformdirs
pluggy
py
py4j
pyparsing
six
toml
tox
virtualenv
apache-airflow==2.7.3
pandas-gbq==0.22.0
openai
